+++ operator/pkg/k3d/integration_on.go
@@ -1,0 +1,16 @@
+// Copyright 2025 Redpanda Data, Inc.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.md
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0
+
+//go:build integration || acceptance
+
+package k3d
+
+func init() {
+	usagePermitted = true
+}
